Refactor "UserService" into "LumaService"
