+++ src/services/mod.rs
@@ -1,1 +1,5 @@
-pub mod user; 
+pub mod luma;
+pub mod auth;
+pub mod users;
+pub mod openai;
+pub mod pinecone;
