+++ src/services/pinecone.rs
