+++ src/services/users/mod.rs
@@ -1,0 +1,2 @@
+pub mod retrieve;
+pub mod delete;
